Rename iescargot to slakje.
Thanks to Sebastian who suggested this historically meaningful name.
